Refactor getCategories finished
