+++ app/src/main/java/ru/investlifestyle/app/domain/repository/PostRepositoryInterface.kt
@@ -2,7 +2,6 @@ package ru.investlifestyle.app.domain.repository
 
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
-import ru.investlifestyle.app.data.models.categories.SaveCategoriesData
 import ru.investlifestyle.app.data.networkApi.PostsModelDataItem
 import ru.investlifestyle.app.ui.models.PostUiModel
 
@@ -32,12 +31,6 @@ interface PostRepositoryInterface {
         perPage: Int,
         embed: Boolean
     ): List<PostUiModel>
-    fun getCategories(): Flow<List<SaveCategoriesData>>
-    fun getCategoriesForChoiceFragment(): Flow<List<SaveCategoriesData>>
-    fun getQuotes(): String
-    suspend fun updateSubject(selected: Boolean, idCategory: Int)
-    suspend fun getSingleSubjectById(idCategories: Int): SaveCategoriesData
-    suspend fun fillingDbInit()
     suspend fun likePostIsEmpty(): Boolean
     suspend fun getLikePostById(postId: Int): PostUiModel
     suspend fun getAllLikePosts(): List<PostUiModel>
