+++ app/src/main/java/ru/investlifestyle/app/domain/usecase/FillingDbInitUseCase.kt
@@ -1,12 +1,12 @@
 package ru.investlifestyle.app.domain.usecase
 
 import javax.inject.Inject
-import ru.investlifestyle.app.domain.repository.PostRepositoryInterface
+import ru.investlifestyle.app.domain.repository.CategoriesRepositoryInterface
 
 class FillingDbInitUseCase @Inject constructor(
-    private val postRepositoryInterface: PostRepositoryInterface
+    private val repositoryInterface: CategoriesRepositoryInterface
 ) {
     suspend fun fillingDbInit() {
-        postRepositoryInterface.fillingDbInit()
+        repositoryInterface.fillingDbInit()
     }
 }
