+++ app/src/main/java/ru/investlifestyle/app/domain/usecase/GetQuotesUseCase.kt
@@ -1,12 +1,12 @@
 package ru.investlifestyle.app.domain.usecase
 
 import javax.inject.Inject
-import ru.investlifestyle.app.domain.repository.PostRepositoryInterface
+import ru.investlifestyle.app.domain.repository.CategoriesRepositoryInterface
 
 class GetQuotesUseCase @Inject constructor(
-    private val postRepository: PostRepositoryInterface
+    private val repository: CategoriesRepositoryInterface
 ) {
     fun getQuotes(): String {
-        return postRepository.getQuotes()
+        return repository.getQuotes()
     }
 }
