--- app/src/main/java/ru/investlifestyle/app/domain/usecase/GetSingleSubjectByIdUseCase.kt
@@ -1,13 +1,0 @@
-package ru.investlifestyle.app.domain.usecase
-
-import javax.inject.Inject
-import ru.investlifestyle.app.data.models.categories.SaveCategoriesData
-import ru.investlifestyle.app.domain.repository.PostRepositoryInterface
-
-class GetSingleSubjectByIdUseCase @Inject constructor(
-    private val postRepositoryInterface: PostRepositoryInterface
-) {
-    suspend fun getSingleSubjectById(id: Int): SaveCategoriesData {
-        return postRepositoryInterface.getSingleSubjectById(id)
-    }
-}
